# feat(hydro_deploy_integration,hydro_lang): allow running generated binaries with single-threaded Tokio runtime