--- a/hydro_deploy/core/src/custom_service.rs
+++ b/hydro_deploy/core/src/custom_service.rs
@@ -105,8 +105,8 @@ impl CustomClientPort {
             .load_instantiated(&|p| p)
             .await
             .instantiate()
-            .connect::<ConnectedDirect>()
             .await
+            .connect::<ConnectedDirect>()
     }
 }