--- a/hydro_deploy/hydro_deploy_examples/examples/tagged_stdout_receiver/main.rs
+++ b/hydro_deploy/hydro_deploy_examples/examples/tagged_stdout_receiver/main.rs
@@ -7,7 +7,6 @@ async fn main() {
     let echo_recv = ports
         .port("echo")
         .connect::<ConnectedTagged<ConnectedDirect>>()
-        .await
         .into_source();
 
     let df = dfir_syntax! {