--- a/hydro_lang/src/deploy/trybuild.rs
+++ b/hydro_lang/src/deploy/trybuild.rs
@@ -164,7 +164,7 @@ pub fn compile_graph_trybuild(
             #dfir_expr
         }
 
-        #[hydro_lang::runtime_support::tokio::main(crate = "hydro_lang::runtime_support::tokio")]
+        #[hydro_lang::runtime_support::tokio::main(crate = "hydro_lang::runtime_support::tokio", flavor = "current_thread")]
         async fn main() {
             let ports = hydro_lang::runtime_support::dfir_rs::util::deploy::init_no_ack_start().await;
             let flow = __hydro_runtime(&ports);