Before, we had a janky architecture for establishing network connections which relied on blocking on futures outside an async context, which required a multi-threaded runtime. Now, we establish all connections before launching the DFIR code, so that no blocking is required there.

Fixes #1807