@shadaj 2025-07-23
Non-scientific benchmark indicates a significant performance increase:

Old:
```
[hydro_test::cluster::paxos_bench::Aggregator (process 3)] Latency p50: 0.446 | p99 0.506 ms | p999 0.617 ms (24631392 samples)
[hydro_test::cluster::paxos_bench::Aggregator (process 3)] Throughput: 649249.45 - 665293.06 - 681336.66 requests/s
```

New:
```
[hydro_test::cluster::paxos_bench::Aggregator (process 3)] Latency p50: 0.378 | p99 0.511 ms | p999 0.867 ms (19489681 samples)
[hydro_test::cluster::paxos_bench::Aggregator (process 3)] Throughput: 775913.07 - 779601.21 - 783289.36 requests/s
```